Get the names of available shapes

Create a data frame of available shapes and associated sets. This may be filtered and used as a vector of strings in scale_shape_manual().

Usage

shapes_cast()

Value

A data frame of available sets and shapes.

Examples

# Returns a data frame of available shapes
shapes_cast()
#>          set      shape
#> 1     circle    circleF
#> 3     circle    circleL
#> 5     circle    circleR
#> 7  container        jar
#> 9  container       tube
#> 11     cross     cross1
#> 13     cross     cross2
#> 15    flower sunflower1
#> 17    flower sunflower2
#> 19    flower sunflower3
#> 21    flower sunflower4
#> 23    flower sunflower5
#> 25    flower sunflower6
#> 27    flower sunflower7
#> 29    flower sunflower8
#> 31      geom        box
#> 33      geom     dendro
#> 35      geom     ribbon
#> 37      geom     violin
#> 39      leaf   hibiscus
#> 41      leaf        oak
#> 43   penguin     adelie
#> 45   penguin  chinstrap
#> 47   penguin     gentoo
#> 49   polygon   heptagon
#> 51   polygon    hexagon
#> 53   polygon    octagon
#> 55   polygon   pentagon